+++ workspace/packages/utils/misc/type.ts
@@ -18,3 +18,19 @@
 export function isString(value: any): value is string {
     return typeof value === "string";
 }
+
+export function isNumber(value: any): value is number {
+    return typeof value === "number";
+}
+
+export function isNull(value: any): value is null {
+    return value === null;
+}
+
+export function isUndefined(value: any): value is undefined {
+    return value === undefined;
+}
+
+export function isNone(value: any): value is null | undefined {
+    return isNull(value) || isUndefined(value);
+}